requirements.txt
asgiref==3.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.4
sqlparse==0.5.1
typing_extensions==4.4.0
tzdata==2024.1